# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "<h2>To be able to bypass the JWT authentication for the endpoints needing authorization,         login first, then use the token recieved, press the <b>Authorize</b> button and type 'Bearer ' and paste the token copied.         After, you can access the enpoints requiring authorization without filling the <b>Authorization</b> field.</h2>"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "paths": {
...
    "/categories/": {
      "get": {
        "tags": [
          "The Category Table Endpoints"
        ],
        "description": "<h3>Gets the category list.</h3>",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "The Category Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, creates a category based on validated parameters provided in the request's body.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/categoryCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "The Category Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, changes a category based on validated parameters provided in the request's body.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/categoryChange"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "The Category Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, removes a category based on validated parameters provided in the request's body. This happens only if there are no products with this category.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/categoryDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/memberships/": {
      "get": {
        "tags": [
          "The Membership Table Endpoints"
        ],
        "description": "<h3>Gets the membership list.</h3>",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "The Membership Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, creates a membership based on validated parameters provided in the request's body.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/membershipCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "The Membership Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, changes a membership based on validated parameters provided in the request's body. This only happens if it is not a default membership.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/membershipChange"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "The Membership Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, removes a membership based on validated parameters provided in the request's body. This happens only if it is not a default membership or there are no users with this membership.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/membershipDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/categories": {
      "get": {
        "tags": [
          "The Admin Front-End Endpoints"
        ],
        "description": "<h3>Renders the categories page if the token owner has an admin role. If not admin then returns the login page back to the user.</h3>",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "categoryCreate": {
      "type": "object",
      "properties": {
        "category": {
          "type": "string",
          "example": "Phones"
        }
      },
      "required": [
        "category"
      ]
    },
    "categoryChange": {
      "type": "object",
      "properties": {
        "categoryid": {
          "type": "number",
          "example": 2
        },
        "category": {
          "type": "string",
          "example": "Phones"
        }
      },
      "required": [
        "categoryid",
        "category"
      ]
    },
    "categoryDelete": {
      "type": "object",
      "properties": {
        "categoryid": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "categoryid"
      ]
    },
...
    "membershipCreate": {
      "type": "object",
      "properties": {
        "membership": {
          "type": "string",
          "example": "Gold"
        }
      },
      "required": [
        "membership"
      ]
    },
    "membershipChange": {
      "type": "object",
      "properties": {
        "membershipid": {
          "type": "number",
          "example": 2
        },
        "membership": {
          "type": "string",
          "example": "Gold"
        },
        "discount": {
          "type": "number",
          "example": 30
        }
      },
      "required": [
        "membershipid",
        "membership",
        "discount"
      ]
    },
    "membershipDelete": {
      "type": "object",
      "properties": {
        "membershipid": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "membershipid"
      ]
    },
...
  }
}